nvim: Remove `null-ls` plug and separate config
