+++ nvim/init.lua
@@ -3,25 +3,28 @@
 -- ├── ...
 -- └── lua
 --     ├── plugins
---     │   ├── init.lua
---     │   └── telescope.lua
+--     │   ├── PLUGIN_MODULE_1.lua
+--     │   ├── PLUGIN_MODULE_2.lua
+--     │   ├── PLUGIN_MODULE_3.lua
+--     │   └── PLUGIN_MODULE_N.lua
 --     └── user
 --         ├── autocmds.lua
 --         ├── lazy_bootstrap.lua
 --         ├── maps.lua
---         └── options.lua
+--         ├── options.lua
+--         └── util.lua
 
 ----------------------------------------------[[ Bootstrap Lazy ]]
 
-require("user.lazy_bootstrap")
+require 'user.lazy_bootstrap'
 
 ----------------------------------------------[[  User Settings ]]
 
-require("user.autocmds")
-require("user.options")
-require("user.maps")
+require 'user.autocmds'
+require 'user.options'
+require 'user.maps'
+require 'user.util'
 
 ----------------------------------------------[[  Load Plugins  ]]
 
-require("lazy").setup("plugins") -- Loads each lua/plugin/*
-
+require('lazy').setup 'plugins' -- Loads each lua/plugin/*
