+++ nvim/plugin/jdtls.lua
@@ -2,357 +2,314 @@ local java_cmds = vim.api.nvim_create_augroup('java_cmds', { clear = true })
 local cache_vars = {}
 
 local root_files = {
-  '.git',
-  'mvnw',
-  'gradlew',
-  'pom.xml',
-  'build.gradle',
+    '.git',
+    'mvnw',
+    'gradlew',
+    'pom.xml',
+    'build.gradle',
 }
 
 local features = {
-  -- change this to `true` to enable codelens
-  codelens = true,
+    -- change this to `true` to enable codelens
+    codelens = true,
 
-  -- change this to `true` if you have `nvim-dap`,
-  -- `java-test` and `java-debug-adapter` installed
-  debugger = true,
+    -- change this to `true` if you have `nvim-dap`,
+    -- `java-test` and `java-debug-adapter` installed
+    debugger = true,
 
-  -- change this to `true` to enable decompiler capabilities
-  decompiler = true,
+    -- change this to `true` to enable decompiler capabilities
+    decompiler = true,
 }
 
 local function get_jdtls_paths()
-  if cache_vars.paths then
-    return cache_vars.paths
-  end
+    if cache_vars.paths then
+        return cache_vars.paths
+    end
 
-  local path = {}
+    local path = {}
 
-  path.data_dir = vim.fn.stdpath 'cache' .. '/nvim-jdtls'
+    path.data_dir = vim.fn.stdpath 'cache' .. '/nvim-jdtls'
 
-  local jdtls_install = require('mason-registry').get_package('jdtls'):get_install_path()
+    local mason_registry = require 'mason-registry'
 
-  path.java_agent = jdtls_install .. '/lombok.jar'
-  path.launcher_jar = vim.fn.glob(jdtls_install .. '/plugins/org.eclipse.equinox.launcher_*.jar')
+    local jdtls_install = mason_registry.get_package('jdtls'):get_install_path()
 
-  if vim.fn.has 'mac' == 1 then
-    path.platform_config = jdtls_install .. '/config_mac'
-  elseif vim.fn.has 'unix' == 1 then
-    path.platform_config = jdtls_install .. '/config_linux'
-  elseif vim.fn.has 'win32' == 1 then
-    path.platform_config = jdtls_install .. '/config_win'
-  end
+    path.java_agent = jdtls_install .. '/lombok.jar'
+    path.launcher_jar = vim.fn.glob(jdtls_install .. '/plugins/org.eclipse.equinox.launcher_*.jar')
 
-  path.bundles = {}
+    if vim.fn.has 'mac' == 1 then
+        path.platform_config = jdtls_install .. '/config_mac'
+    elseif vim.fn.has 'unix' == 1 then
+        path.platform_config = jdtls_install .. '/config_linux'
+    elseif vim.fn.has 'win32' == 1 then
+        path.platform_config = jdtls_install .. '/config_win'
+    end
 
-  ---
-  -- Include java-test bundle if present
-  ---
-  local java_test_path = require('mason-registry').get_package('java-test'):get_install_path()
+    path.bundles = {}
 
-  local java_test_bundle = vim.split(vim.fn.glob(java_test_path .. '/extension/server/*.jar'), '\n')
+    ---
+    -- Include java-test bundle if present
+    ---
+    local java_test_path = mason_registry.get_package('java-test'):get_install_path()
 
-  if java_test_bundle[1] ~= '' then
-    vim.list_extend(path.bundles, java_test_bundle)
-  end
+    local java_test_bundle =
+        vim.split(vim.fn.glob(java_test_path .. '/extension/server/*.jar'), '\n')
 
-  ---
-  -- Include java-debug-adapter bundle if present
-  ---
-  local java_debug_path =
-    require('mason-registry').get_package('java-debug-adapter'):get_install_path()
+    if java_test_bundle[1] ~= '' then
+        vim.list_extend(path.bundles, java_test_bundle)
+    end
 
-  local java_debug_bundle = vim.split(
-    vim.fn.glob(java_debug_path .. '/extension/server/com.microsoft.java.debug.plugin-*.jar'),
-    '\n'
-  )
+    ---
+    -- Include java-debug-adapter bundle if present
+    ---
+    local java_debug_path = mason_registry.get_package('java-debug-adapter'):get_install_path()
 
-  if java_debug_bundle[1] ~= '' then
-    vim.list_extend(path.bundles, java_debug_bundle)
-  end
+    local java_debug_bundle = vim.split(
+        vim.fn.glob(java_debug_path .. '/extension/server/com.microsoft.java.debug.plugin-*.jar'),
+        '\n'
+    )
 
-  ---
-  -- Include java-decompiler bundle if present
-  ---
-  local java_decompiler_jars =
-    vim.fn.glob(vim.fn.expand '$HOME' .. '/.local/source/vscode-java-decompiler/server/*.jar')
+    if java_debug_bundle[1] ~= '' then
+        vim.list_extend(path.bundles, java_debug_bundle)
+    end
 
-  local java_decompiler_bundle = vim.fn.split(java_decompiler_jars, '\n')
+    ---
+    -- Include java-decompiler bundle if present
+    ---
+    local java_decompiler_path =
+        mason_registry.get_package('vscode-java-decompiler'):get_install_path()
+    local java_decompiler_jars = vim.fn.glob(java_decompiler_path .. '/server/*.jar')
 
-  if features.decompiler then
-    vim.list_extend(path.bundles, java_decompiler_bundle)
-  end
+    local java_decompiler_bundle = vim.fn.split(java_decompiler_jars, '\n')
 
-  ---
-  -- Useful if you're starting jdtls with a Java version that's
-  -- different from the one the project uses.
-  ---
-  path.runtimes = {
-    -- Note: the field `name` must be a valid `ExecutionEnvironment`,
-    -- you can find the list here:
-    -- https://github.com/eclipse/eclipse.jdt.ls/wiki/Running-the-JAVA-LS-server-from-the-command-line#initialize-request
-    --
-    -- This example assume you are using sdkman: https://sdkman.io
-    -- {
-    --   name = 'JavaSE-17',
-    --   path = vim.fn.expand('~/.sdkman/candidates/java/17.0.6-tem'),
-    -- },
-    -- {
-    --   name = 'JavaSE-18',
-    --   path = vim.fn.expand('~/.sdkman/candidates/java/18.0.2-amzn'),
-    -- },
-  }
+    if features.decompiler then
+        vim.list_extend(path.bundles, java_decompiler_bundle)
+    end
 
-  cache_vars.paths = path
+    ---
+    -- Useful if you're starting jdtls with a Java version that's
+    -- different from the one the project uses.
+    ---
+    path.runtimes = {
+        -- Note: the field `name` must be a valid `ExecutionEnvironment`,
+        -- you can find the list here:
+        -- https://github.com/eclipse/eclipse.jdt.ls/wiki/Running-the-JAVA-LS-server-from-the-command-line#initialize-request
+        --
+        -- This example assume you are using sdkman: https://sdkman.io
+        -- {
+        --   name = 'JavaSE-17',
+        --   path = vim.fn.expand('~/.sdkman/candidates/java/17.0.6-tem'),
+        -- },
+        -- {
+        --   name = 'JavaSE-18',
+        --   path = vim.fn.expand('~/.sdkman/candidates/java/18.0.2-amzn'),
+        -- },
+    }
 
-  return path
+    cache_vars.paths = path
+
+    return path
 end
 
 local function enable_codelens(bufnr)
-  pcall(vim.lsp.codelens.refresh)
+    pcall(vim.lsp.codelens.refresh)
 
-  vim.api.nvim_create_autocmd('BufWritePost', {
-    buffer = bufnr,
-    group = java_cmds,
-    desc = 'refresh codelens',
-    callback = function()
-      pcall(vim.lsp.codelens.refresh)
-    end,
-  })
+    vim.api.nvim_create_autocmd('BufWritePost', {
+        buffer = bufnr,
+        group = java_cmds,
+        desc = 'refresh codelens',
+        callback = function()
+            pcall(vim.lsp.codelens.refresh)
+        end,
+    })
 end
 
 local function attach_to_debug()
-  local dap = require 'dap'
+    local dap = require 'dap'
 
-  dap.configurations.java = {
-    {
-      type = 'java',
-      request = 'attach',
-      name = 'Attach to the process',
-      hostName = 'localhost',
-      port = '5005',
-    },
-  }
+    dap.configurations.java = {
+        {
+            type = 'java',
+            request = 'attach',
+            name = 'Attach to the process',
+            hostName = 'localhost',
+            port = '5005',
+        },
+    }
 
-  dap.continue()
+    dap.continue()
 end
 
 local function enable_debugger(bufnr)
-  require('jdtls').setup_dap { hotcodereplace = 'auto' }
-  require('jdtls.dap').setup_dap_main_class_configs()
+    local jdtls = require 'jdtls'
 
-  local function create_opts(description)
-    return { buffer = bufnr, desc = description }
-  end
+    jdtls.setup_dap { hotcodereplace = 'auto' }
+    require('jdtls.dap').setup_dap_main_class_configs()
 
-  vim.keymap.set(
-    'n',
-    '<leader>dtc',
-    "<cmd>lua require('jdtls').test_class()<cr>",
-    create_opts 'Debug Test'
-  )
-  vim.keymap.set(
-    'n',
-    '<leader>dtn',
-    "<cmd>lua require('jdtls').test_nearest_method()<cr>",
-    create_opts 'Debug Nearest Test'
-  )
-  vim.keymap.set('n', '<leader>dAj', attach_to_debug, create_opts 'Debug Last Test')
+    local function create_opts(description)
+        return { buffer = bufnr, desc = description }
+    end
+
+    vim.keymap.set('n', '<leader>dtc', jdtls.test_class, create_opts 'Debug Class')
+    vim.keymap.set('n', '<leader>dtn', jdtls.test_nearest_method, create_opts 'Debug Nearest Test')
+    vim.keymap.set('n', '<leader>da', attach_to_debug, create_opts 'Attach to Debugging Session')
 end
 
 local function jdtls_on_attach(client, bufnr)
-  if features.debugger then
-    enable_debugger(bufnr)
-  end
+    if features.debugger then
+        enable_debugger(bufnr)
+    end
 
-  if features.codelens then
-    enable_codelens(bufnr)
-  end
+    if features.codelens then
+        enable_codelens(bufnr)
+    end
 
-  -- The following mappings are based on the suggested usage of nvim-jdtls
-  -- https://github.com/mfussenegger/nvim-jdtls#usage
+    -- The following mappings are based on the suggested usage of nvim-jdtls
+    -- https://github.com/mfussenegger/nvim-jdtls#usage
 
-  local function create_opts(description)
-    return { buffer = bufnr, desc = description }
-  end
-
-  vim.keymap.set(
-    'n',
-    '<A-o>',
-    "<cmd>lua require('jdtls').organize_imports()<cr>",
-    create_opts 'Organize Imports'
-  )
-  vim.keymap.set(
-    'n',
-    'crv',
-    "<cmd>lua require('jdtls').extract_variable()<cr>",
-    create_opts 'Extract Variable'
-  )
-  vim.keymap.set(
-    'x',
-    'crv',
-    "<esc><cmd>lua require('jdtls').extract_variable(true)<cr>",
-    create_opts 'Extract Variable'
-  )
-  vim.keymap.set(
-    'n',
-    'crc',
-    "<cmd>lua require('jdtls').extract_constant()<cr>",
-    create_opts 'Extract Constant'
-  )
-  vim.keymap.set(
-    'x',
-    'crc',
-    "<esc><cmd>lua require('jdtls').extract_constant(true)<cr>",
-    create_opts 'Extract Constant'
-  )
-  vim.keymap.set(
-    'x',
-    'crm',
-    "<esc><Cmd>lua require('jdtls').extract_method(true)<cr>",
-    create_opts 'Extract Method'
-  )
+    -- Add general LSP keybindings
+    require('user.util').on_attach(client, bufnr)
 end
 
-local function jdtls_setup(event)
-  local jdtls = require 'jdtls'
+local function jdtls_setup()
+    local jdtls = require 'jdtls'
 
-  local path = get_jdtls_paths()
-  local data_dir = path.data_dir .. '/' .. vim.fn.fnamemodify(vim.fn.getcwd(), ':p:h:t')
+    local path = get_jdtls_paths()
+    local data_dir = path.data_dir .. '/' .. vim.fn.fnamemodify(vim.fn.getcwd(), ':p:h:t')
 
-  if cache_vars.capabilities == nil then
-    jdtls.extendedClientCapabilities.resolveAdditionalTextEditsSupport = true
+    if cache_vars.capabilities == nil then
+        jdtls.extendedClientCapabilities.resolveAdditionalTextEditsSupport = true
 
-    local ok_cmp, cmp_lsp = pcall(require, 'cmp_nvim_lsp')
-    cache_vars.capabilities = vim.tbl_deep_extend(
-      'force',
-      vim.lsp.protocol.make_client_capabilities(),
-      ok_cmp and cmp_lsp.default_capabilities() or {}
-    )
-  end
+        local ok_cmp, cmp_lsp = pcall(require, 'cmp_nvim_lsp')
+        cache_vars.capabilities = vim.tbl_deep_extend(
+            'force',
+            vim.lsp.protocol.make_client_capabilities(),
+            ok_cmp and cmp_lsp.default_capabilities() or {}
+        )
+    end
 
-  -- The command that starts the language server
-  -- See: https://github.com/eclipse/eclipse.jdt.ls#running-from-the-command-line
-  local cmd = {
-    -- 💀
-    'java',
+    -- The command that starts the language server
+    -- See: https://github.com/eclipse/eclipse.jdt.ls#running-from-the-command-line
+    local cmd = {
+        -- 💀
+        'java',
 
-    '-Declipse.application=org.eclipse.jdt.ls.core.id1',
-    '-Dosgi.bundles.defaultStartLevel=4',
-    '-Declipse.product=org.eclipse.jdt.ls.core.product',
-    '-Dlog.protocol=true',
-    '-Dlog.level=ALL',
-    '-javaagent:' .. path.java_agent,
-    '-Xms1g',
-    '--add-modules=ALL-SYSTEM',
-    '--add-opens',
-    'java.base/java.util=ALL-UNNAMED',
-    '--add-opens',
-    'java.base/java.lang=ALL-UNNAMED',
+        '-Declipse.application=org.eclipse.jdt.ls.core.id1',
+        '-Dosgi.bundles.defaultStartLevel=4',
+        '-Declipse.product=org.eclipse.jdt.ls.core.product',
+        '-Dlog.protocol=true',
+        '-Dlog.level=ALL',
+        '-javaagent:' .. path.java_agent,
+        '-Xms1g',
+        '--add-modules=ALL-SYSTEM',
+        '--add-opens',
+        'java.base/java.util=ALL-UNNAMED',
+        '--add-opens',
+        'java.base/java.lang=ALL-UNNAMED',
 
-    -- 💀
-    '-jar',
-    path.launcher_jar,
+        -- 💀
+        '-jar',
+        path.launcher_jar,
 
-    -- 💀
-    '-configuration',
-    path.platform_config,
+        -- 💀
+        '-configuration',
+        path.platform_config,
 
-    -- 💀
-    '-data',
-    data_dir,
-  }
+        -- 💀
+        '-data',
+        data_dir,
+    }
 
-  local lsp_settings = {
-    java = {
-      -- jdt = {
-      --   ls = {
-      --     vmargs = "-XX:+UseParallelGC -XX:GCTimeRatio=4 -XX:AdaptiveSizePolicyWeight=90 -Dsun.zip.disableMemoryMapping=true -Xmx1G -Xms100m"
-      --   }
-      -- },
-      eclipse = {
-        downloadSources = true,
-      },
-      configuration = {
-        updateBuildConfiguration = 'interactive',
-        runtimes = path.runtimes,
-      },
-      maven = {
-        downloadSources = true,
-      },
-      implementationsCodeLens = {
-        enabled = true,
-      },
-      referencesCodeLens = {
-        enabled = true,
-      },
-      references = {
-        includeDecompiledSources = true,
-      },
-      inlayHints = {
-        parameterNames = {
-          enabled = 'all', -- literals, all, none
+    local lsp_settings = {
+        java = {
+            -- jdt = {
+            --   ls = {
+            --     vmargs = "-XX:+UseParallelGC -XX:GCTimeRatio=4 -XX:AdaptiveSizePolicyWeight=90 -Dsun.zip.disableMemoryMapping=true -Xmx1G -Xms100m"
+            --   }
+            -- },
+            eclipse = {
+                downloadSources = true,
+            },
+            configuration = {
+                updateBuildConfiguration = 'interactive',
+                runtimes = path.runtimes,
+            },
+            maven = {
+                downloadSources = true,
+            },
+            implementationsCodeLens = {
+                enabled = true,
+            },
+            referencesCodeLens = {
+                enabled = true,
+            },
+            references = {
+                includeDecompiledSources = true,
+            },
+            inlayHints = {
+                parameterNames = {
+                    enabled = 'all', -- literals, all, none
+                },
+            },
+            format = {
+                enabled = true,
+                -- settings = {
+                --   profile = 'asdf'
+                -- },
+            },
         },
-      },
-      format = {
-        enabled = true,
-        -- settings = {
-        --   profile = 'asdf'
-        -- },
-      },
-    },
-    signatureHelp = {
-      enabled = true,
-    },
-    completion = {
-      favoriteStaticMembers = {
-        'org.hamcrest.MatcherAssert.assertThat',
-        'org.hamcrest.Matchers.*',
-        'org.hamcrest.CoreMatchers.*',
-        'org.junit.jupiter.api.Assertions.*',
-        'java.util.Objects.requireNonNull',
-        'java.util.Objects.requireNonNullElse',
-        'org.mockito.Mockito.*',
-      },
-    },
-    contentProvider = {
-      preferred = 'fernflower',
-    },
-    extendedClientCapabilities = jdtls.extendedClientCapabilities,
-    sources = {
-      organizeImports = {
-        starThreshold = 9999,
-        staticStarThreshold = 9999,
-      },
-    },
-    codeGeneration = {
-      toString = {
-        template = '${object.className}{${member.name()}=${member.value}, ${otherMembers}}',
-      },
-      useBlocks = true,
-    },
-  }
+        signatureHelp = {
+            enabled = true,
+        },
+        completion = {
+            favoriteStaticMembers = {
+                'org.hamcrest.MatcherAssert.assertThat',
+                'org.hamcrest.Matchers.*',
+                'org.hamcrest.CoreMatchers.*',
+                'org.junit.jupiter.api.Assertions.*',
+                'java.util.Objects.requireNonNull',
+                'java.util.Objects.requireNonNullElse',
+                'org.mockito.Mockito.*',
+            },
+        },
+        contentProvider = {
+            preferred = 'fernflower',
+        },
+        extendedClientCapabilities = jdtls.extendedClientCapabilities,
+        sources = {
+            organizeImports = {
+                starThreshold = 9999,
+                staticStarThreshold = 9999,
+            },
+        },
+        codeGeneration = {
+            toString = {
+                template = '${object.className}{${member.name()}=${member.value}, ${otherMembers}}',
+            },
+            useBlocks = true,
+        },
+    }
 
-  -- This starts a new client & server,
-  -- or attaches to an existing client & server depending on the `root_dir`.
-  jdtls.start_or_attach {
-    cmd = cmd,
-    settings = lsp_settings,
-    on_attach = jdtls_on_attach,
-    capabilities = cache_vars.capabilities,
-    root_dir = jdtls.setup.find_root(root_files),
-    flags = {
-      allow_incremental_sync = true,
-    },
-    init_options = {
-      bundles = path.bundles,
-    },
-  }
+    -- This starts a new client & server,
+    -- or attaches to an existing client & server depending on the `root_dir`.
+    jdtls.start_or_attach {
+        cmd = cmd,
+        settings = lsp_settings,
+        on_attach = jdtls_on_attach,
+        capabilities = cache_vars.capabilities,
+        root_dir = jdtls.setup.find_root(root_files),
+        flags = {
+            allow_incremental_sync = true,
+        },
+        init_options = {
+            bundles = path.bundles,
+        },
+    }
 end
 
 vim.api.nvim_create_autocmd('FileType', {
-  group = java_cmds,
-  pattern = { 'java' },
-  desc = 'Setup jdtls',
-  callback = jdtls_setup,
+    group = java_cmds,
+    pattern = { 'java' },
+    desc = 'Setup JDTLS',
+    callback = jdtls_setup,
 })
